pyproject.toml
[project]
name = "mcp-weather-server"
version = "0.2.0"
description = "Refactored HTTP Streamable based MCP Server"
requires-python = ">=3.12"
dependencies = [
    "httpx~=0.28.0",
    "mcp[cli]~=1.2.0",  # mcp 버전에 맞게 조정
    "uvicorn",
    "python-dotenv"     # 추가됨
]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"